move pruning to store
