+++ server/pruning.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/spf13/cast"
 
-	pruningtypes "github.com/cosmos/cosmos-sdk/pruning/types"
 	"github.com/cosmos/cosmos-sdk/server/types"
+	pruningtypes "github.com/cosmos/cosmos-sdk/store/pruning/types"
 )
 
 // GetPruningOptionsFromFlags parses command flags and returns the correct
