+++ store/pruning/README.md
